scripts/requirements.txt
pendulum
python-dotenv
